fix for issue 169, on some windows 7
On some windows machine alerts are shown, when something goes bad when
trying to list com ports. dhToggleExceptions(FALSE) turns those off.
But then, there is sometimes huge number of serial ports reported, which do not have their
name set. We check for the name existence and proceed only when it is
available. For us it fixes
https://github.com/voodootikigod/node-serialport/issues/169
reliably on multiple machines.